// apps/collector-fb/src/types/enums.ts
export type FunnelStage = "top" | "bottom";
export type FacebookEventType =
  | "ad_view"
  | "page_like"
  | "comment"
  | "video_view"
  | "ad_click"
  | "form_submission"
  | "checkout_complete";
export type Gender = "male" | "female" | "non_binary";
export type Referrer = "newsfeed" | "marketplace" | "groups";
export type ClickPosition = "top_left" | "bottom_right" | "center";
export type Device = "mobile" | "desktop";
export type Browser = "Chrome" | "Firefox" | "Safari";